Crate medschool

medschool will extract the docs from the current directory, and build a nice configuration.md file.

To use it simply run medschool on the directory you want to get docs from, for example:

cd rust-project
medschool

It’ll look into rust-project/src and produce rust-project/configuration.md.

Structs

PartialField 🔒

The fields when the item we’re looking is syn::ItemStruct.

PartialType 🔒

We’re extracting syn::Item structs and enums into this type.

Enums

DocsError 🔒

We just eat some of these errors (turn them into None).

Functions

docs_from_attributes 🔒

Look into the syn::Attributes of whatever item we’re handling, and extract its doc strings.

get_ident_from_field_skipping_generics 🔒

Digs into the generics of a field (syn::ItemStruct only), trying to get the last syn::PathArguments, which (hopefully) contains the concrete type we care about.

main 🔒

parse_files 🔒

Parses all files in the glob::glob pattern defined within, in the current directory.

pretty_docs 🔒

Glues all the Vec<String> docs into one big String.