Crate medschool

medschool will extract the docs from the current directory, and build a nice configuration.md file.

To use it simply run medschool on the directory you want to get docs from, for example:

cd rust-project
medschool

It’ll look into rust-project/src and produce rust-project/configuration.md.

Structs

PartialField 🔒

The fields when the item we’re looking is syn::ItemStruct.

PartialType 🔒

We’re extracting syn::Item structs and enums into this type.

Enums

DocsError 🔒

We just eat some of these errors (turn them into None).

Statics

CALLSITE 🔒

CALLSITE 🔒

Functions

depth_first_build_new_types 🔒

Digs into the [PartialTypes] building new types that inline the types of their PartialFields, turning something like:

docs_from_attributes 🔒

Look into the syn::Attributes of whatever item we’re handling, and extract its doc strings.

files_to_string 🔒

Converts all files in the glob::glob pattern defined within, in the current directory, into a Vec<String>.

get_ident_from_field_skipping_generics 🔒

Digs into the generics of a field (syn::ItemStruct only), trying to get the last syn::PathArguments, which (hopefully) contains the concrete type we care about.

get_root_type 🔒

Gets the element with the most number of PartialField, which at this point should be our root PartialType.

main 🔒

Attention when using RUST_LOG

parse_docs_into_tree 🔒

Converts a list of syn::File into a BTreeSet of our own PartialType types, so we can get a root node (see the Ord implementation of PartialType).

parse_string_files 🔒

Parses the files into a collection of syn::File.

pretty_docs 🔒

Glues all the Vec<String> docs into one big String.

produce_docs_from_root_type 🔒

Turns the root PartialType documentation into one big String.